ClimFlowsTestCases

Documentation for ClimFlowsTestCases.

ClimFlowsTestCases.DCMIP

ClimFlowsTestCases.Isothermal

ClimFlowsTestCases.Jablonowski06

ClimFlowsTestCases.TestCase

ClimFlowsTestCases.TestCaseHPE

ClimFlowsTestCases.TestCaseSW

ClimFlowsTestCases.Williamson91

ClimFlowsTestCases.describe

ClimFlowsTestCases.initial

ClimFlowsTestCases.DCMIP — Type

struct DCMIP{N} <: TestCaseHPE

DCMIP 2012 test cases. Currently only N=42 (test case 4.2) is implemented.

ClimFlowsTestCases.Isothermal — Type

struct Isothermal <: TestCaseHPE

Return testcase corresponding to an initial state of rest with exponential decrease of pressure with geopotential. For a perfect gas, this coincides with the classic isothermal profile, and psurf * vsurf = R*T with T the constant temperature. The surface geopotential defines a wiggly mountain. Numerics based on terrain-following coordinates typically struggle with such case and generate spurious motion.

ClimFlowsTestCases.Jablonowski06 — Type

struct Jablonowski06{P} <: TestCaseHPE

Jablonowski (2006) and DCMIP 4.2 test cases (dry vs. moist). The only difference is specific humidity is included for the moist case.

ClimFlowsTestCases.TestCase — Type

abstract type TestCase end

Parent type for test case types. Derived types must implement a constructor accepting an optional precision and keyword arguments to override default ones. For example:

case = Williams91{6}(Float32; R0=3e-3)

Derived types must also possess a field params which is a named tuple containing the parameters, and implement initial and describe.

ClimFlowsTestCases.TestCaseHPE — Type

Parent type for hydrostatic test cases.

ClimFlowsTestCases.TestCaseSW — Type

Parent type for shallow-water test cases.

ClimFlowsTestCases.Williamson91 — Type

struct Williamson91{N} <: TestCaseSW

Williamson (1991) test cases. Currently only N=6 is implemented.

ClimFlowsTestCases.describe — Function

string = describe(case::TestCase)

Plain text description of case, possibly multi-line.

ClimFlowsTestCases.initial — Function

gh, ulon, ulat = initial(case::TestCaseSW, lon, lat)

For a shallow-water test case, returns the initial geopotential thickness and velocity at given latitude and longitude.

ps, Phis = initial(case::TestCaseHPE, lon, lat)

For a hydrostatic test case, returns the initial surface pressure and geopotential at given latitude and longitude.

Phi, ulon, ulat, q = initial(case::TestCaseHPE, lon, lat, p)

For a hydrostatic test case, returns the initial geopotential, velocity and composition at given latitude, longitude and pressure.