ClimFlowsTestCases

Documentation for ClimFlowsTestCases.

ClimFlowsTestCases.DCMIP

ClimFlowsTestCases.Jablonowski06

ClimFlowsTestCases.TestCase

ClimFlowsTestCases.TestCaseHPE

ClimFlowsTestCases.TestCaseSW

ClimFlowsTestCases.Williamson91

ClimFlowsTestCases.default_params

ClimFlowsTestCases.describe

ClimFlowsTestCases.initial_flow

ClimFlowsTestCases.initial_surface

ClimFlowsTestCases.testcase

ClimFlowsTestCases.DCMIP — Type

struct DCMIP{N} <: TestCaseHPE

DCMIP 2012 test cases. Currently only N=42 (test case 4.2) is implemented.

ClimFlowsTestCases.Jablonowski06 — Type

struct Jablonowski06{P} <: TestCaseHPE

Jablonowski (2006) and DCMIP 4.2 test cases (dry vs. moist). The only difference is specific humidity is included for the moist case.

ClimFlowsTestCases.TestCase — Type

abstract type TestCase end

Parent type for test case types. Derived type must implement a constructor accepting a named tuple of parameters and the function default_params. For example:

TC = Williams91{6}
params = default_params(TC)
case = TC((params..., R0=3e-3))

ClimFlowsTestCases.TestCaseHPE — Type

Parent type for hydrostatic test cases.

ClimFlowsTestCases.TestCaseSW — Type

Parent type for shallow-water test cases.

ClimFlowsTestCases.Williamson91 — Type

struct Williamson91{N} <: TestCaseSW

Williamson (1991) test cases. Currently only N=6 is implemented.

ClimFlowsTestCases.default_params — Method

params = default_params(TestCaseType)

Returns a named tuple of default parameters for a certain test case. More about named tuples here.

ClimFlowsTestCases.describe — Function

string = describe(case::TestCase)

Plain text description of case, possibly multi-line.

ClimFlowsTestCases.initial_flow — Function

gh, ulon, ulat = initial_flow(lon, lat, case::TestCaseSW)

For a shallow-water test case, returns the initial geopotential thickness and velocity at given latitude and longitude.

Phi, ulon, ulat, q = initial_flow(lon, lat, p, case::TestCaseHPE)

For a hydrostatic test case, returns the initial geopotential, velocity and composition at given latitude, longitude and pressure.

ClimFlowsTestCases.initial_surface — Function

ps, Phis = initial_surface(lon, lat, case::TestCaseHPE)

For a hydrostatic test case, returns the initial surface pressure and geopotential at given latitude and longitude.

ClimFlowsTestCases.testcase — Method

case = testcase(Case::Type, F=Float64 ; params...)

Returns a test case with default parameters for case Case with floating-point precision F. Optional params override the default parameters for the test case. For example:

w91_6 = testcase(Williamson91{6}) # Float64, default parameters
w91_6 = testcase(Williamson91{6}, Float32 ; R0=1e-3) # Float32, overrides R0 (planetary radius)