ClimFlowsTestCases

Documentation for ClimFlowsTestCases.

ClimFlowsTestCases.priv

ClimFlowsTestCases.TestCase

ClimFlowsTestCases.TestCaseHPE

ClimFlowsTestCases.TestCaseSW

ClimFlowsTestCases.priv.DCMIP

ClimFlowsTestCases.priv.Jablonowski06

ClimFlowsTestCases.priv.Williamson91

ClimFlowsTestCases.default_params

ClimFlowsTestCases.TestCase — Type

abstract type TestCase end

Parent type for test case types.

ClimFlowsTestCases.TestCaseHPE — Type

Parent type for hydrostatic test cases.

ClimFlowsTestCases.TestCaseSW — Type

Parent type for shallow-water test cases.

ClimFlowsTestCases.default_params — Method

params = default_params(TestCaseType)

Returns a named tuple of default parameters for a certain test case. More about named tuples here.

ClimFlowsTestCases.priv — Module

Implementation details that may change between non-breaking versions.

ClimFlowsTestCases.priv.DCMIP — Type

struct DCMIP{N} <: TestCaseHPE

Type for DCMIP 2012 test cases. Currently only N=42 (test case 4.2) is implemented.

ClimFlowsTestCases.priv.Jablonowski06 — Type

Jablonowski06

Type for Jablonowski (2006) and DCMIP 4.2 test cases (dry vs. moist). The only difference is specific humidity is included for the moist case.

ClimFlowsTestCases.priv.Williamson91 — Type

Williamson91{N}

Type for Williamson (1991) test cases. Currently only N=6 is implemented.